pyproject.toml
[build-system]
requires = ['setuptools>=69.5.1']
build-backend = "setuptools.build_meta"

[project]
name = "sbxt-accounts"
version = "0.0.0"
description = "standard Sandboxed Thoughts Django users"
readme = "README.rst"
requires-python = ">3.10"
keywords = ["django", "sbxt", "sandboxed thoughts", "users", "apps"]
authors = [
    {name = "Sandboxed Thoughts", email = "info@sandboxedthoughts.com"},
    {name = "Nate Schram", email = "nate@sandboxedthoughts.com"}
]
maintainers = [
    {name = "Nate Schram", email = "nate@sandboxedthoughts.com"}
]
classifiers = [
    # How mature is this project? Common values are
    #   3 - Alpha
    #   4 - Beta
    #   5 - Production/Stable
    "Development Status :: 3 - Alpha"
]
dependencies = [
    "django~=5.0",
    "python-dateutil~=2.9",
    "django-phonenumber-field~=7.3"
]
